fix(server-core): repoint the snapshot tools at the consolidated loader
`main` does not compile. #881 and #884 imported
`../render/load-spatial-canvas.js`, which #886 had already deleted when it
consolidated every document load into `tools/document-io.ts`. Neither PR
conflicted textually and both were green on their own base, so the squash
merge landed a broken tree:

  src/tools/canvas-snapshot.ts:12      TS2307: Cannot find module
  src/tools/canvas-snapshot.test.ts:10 TS2307: Cannot find module
  src/tools/canvas-edit.test.ts:10     TS2307: Cannot find module

`loadDocument` returns the same `{ doc, canvas }` shape the deleted
`loadSpatialCanvas` did, and `assertSpatialDocument` still lives in
`render/`, so this is a pure repointing with no behaviour change.
